90% finalizado

+++ src/resources/material-resources.ts
@@ -4,6 +4,7 @@ import {
   DELETE,
   FileParam,
   GET,
+  HeaderParam,
   Path,
   PathParam,
   POST,
@@ -11,40 +12,70 @@ import {
 } from "typescript-rest";
 import { Material } from "../models/Material";
 import { MaterialService } from "../services/imp/material-service";
-
-import express = require("express");
 import { ExportFileService } from "../services/imp/export-file-service";
 import { AxiosError } from "axios";
+import { JwtMiddleware } from "../middlewares/check-jwt-middleware";
 
+import express = require("express");
 @Path("/v1/material")
 export class MaterialResources {
   @Inject
   private service: MaterialService;
 
   @Inject
+  private authMiddleware: JwtMiddleware;
+
+  @Inject
   private exportFileService: ExportFileService
 
   @GET
   @Path("")
-  public findAll(
+  public async findAll(
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
   ): Promise<any> {
-    return this.service.findAll();
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.findAll();
   }
 
   @GET
   @Path(":uuid")
-  public findByUuid(
+  public async findByUuid(
     @PathParam("uuid") uuid: string,
-  ): Promise<Material> {
-    return this.service.findByUuid(uuid);
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
+  ): Promise<any> {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.findByUuid(uuid);
   }
 
   @POST
   @Path("")
-  public save(
+  public async save(
     material: Material,
-  ): Promise<Material> {
-    return this.service.save(material);
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
+  ): Promise<any> {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.save(material);
   }
 
   @PUT
@@ -52,32 +83,52 @@ export class MaterialResources {
   public async update(
     @PathParam("uuid") uuid: string,
     material: Material,
+    @HeaderParam("authorization") authorization: string,
     @ContextResponse res: express.Response
   ) {
-    try {
-      return this.service.update(uuid, material);
-    } catch (err) {
-      return res.status(400).send("deu ruim");
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
     }
+
+    return await this.service.update(uuid, material);
   }
 
-  // Esse método está desativado, por falta de informações na tabela de material. Obs: "Excluir um dado do banco NÃO É UMA BOA PRÁTICA!".
+  // Esse método não está funcionando por falta de informações na tabela de material. Obs: "Excluir um dado do banco NÃO É UMA BOA PRÁTICA!".
   @DELETE
   @Path(":uuid")
   public delete(
-    @PathParam("uuid") uuid: string
+    @PathParam("uuid") uuid: string,
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
   ) {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
     return this.service.delete(uuid);
   }
 
   @Path('upload')
   @POST
-  public upload(
+  public async upload(
     @FileParam('xls') xls: Express.Multer.File,
+    @HeaderParam("authorization") authorization: string,
     @ContextResponse res: express.Response
-  ): any {
+  ): Promise<any> {
 
-    return this.exportFileService
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.exportFileService
       .uploadMaterial(xls)
       .then(response => {
         return response;
