+++ src/resources/medicine-resources.ts
@@ -3,45 +3,76 @@ import {
   ContextResponse,
   DELETE,
   GET,
+  HeaderParam,
   Path,
   PathParam,
   POST,
   PUT,
 } from "typescript-rest";
 import { MedicineService } from "../services/imp/medicine-service";
+import { Medicines } from "../models/Medicines";
+import { JwtMiddleware } from "../middlewares/check-jwt-middleware";
 
 import express = require("express");
-import { Medicines } from "../models/Medicines";
-
 @Path("/v1/medicines")
 export class MedicineResources {
   @Inject
   private service: MedicineService;
+
+  @Inject
+  private authMiddleware: JwtMiddleware;
 
   /*@Inject
   private exportFileService: ExportFileService*/
 
   @GET
   @Path("")
-  public findAll(
+  public async findAll(
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
   ): Promise<any> {
-    return this.service.findAll();
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.findAll();
   }
 
   @GET
   @Path(":uuid")
-  public findByUuid(
+  public async findByUuid(
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response,
     @PathParam("uuid") uuid: string,
-  ): Promise<Medicines> {
-    return this.service.findByUuid(uuid);
+  ): Promise<any> {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.findByUuid(uuid);
   }
 
   @POST
   @Path("")
-  public save(
+  public async save(
     material: Medicines,
-  ): Promise<Medicines> {
-    return this.service.save(material);
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
+  ): Promise<any> {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
+    return await this.service.save(material);
   }
 
   @PUT
@@ -49,35 +80,48 @@ export class MedicineResources {
   public async update(
     @PathParam("uuid") uuid: string,
     medicine: Medicines,
+    @HeaderParam("authorization") authorization: string,
     @ContextResponse res: express.Response
   ) {
-    try {
-      return this.service.update(uuid, medicine);
-    } catch (err) {
-      return res.status(400).send("deu ruim");
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
     }
+
+    return await this.service.update(uuid, medicine);
   }
 
-  // Esse método está desativado, por falta de informações na tabela de material. Obs: "Excluir um dado do banco NÃO É UMA BOA PRÁTICA!".
+  // Esse método não está funcionando por falta de informações na tabela de material. Obs: "Excluir um dado do banco NÃO É UMA BOA PRÁTICA!".
   @DELETE
   @Path(":uuid")
   public delete(
-    @PathParam("uuid") uuid: string
+    @PathParam("uuid") uuid: string,
+    @HeaderParam("authorization") authorization: string,
+    @ContextResponse res: express.Response
   ) {
+
+    if (!this.authMiddleware.validateAuth(authorization)) {
+      return res.status(400).send({
+        message: "Token inválido"
+      });
+    }
+
     return this.service.delete(uuid);
   }
 
   /*@Path('upload')
   @POST
-  public upload(
+  public async upload(
     @FileParam('xls') xls: Express.Multer.File,
     @ContextResponse res: express.Response
   ): any {
 
-    return this.exportFileService
+    return await this.exportFileService
       .uploadMaterial(xls)
       .then(response => {
-        return response;
+        return await response;
       })
       .catch((e: AxiosError) => {
         console.log('Deu erro na chamada da API', e);
